Adding comments to cupy install and some handling of other distros

+++ rocm/sources/scripts/lmod_setup.sh
@@ -4,8 +4,12 @@ echo ""
 echo "############# Lmod Setup script ################"
 echo ""
 
-sudo DEBIAN_FRONTEND=noninteractive apt-get -qq update
-sudo DEBIAN_FRONTEND=noninteractive apt-get -qqy install lmod
+if [ "${DISTRO}" = "ubuntu" ]; then
+   sudo DEBIAN_FRONTEND=noninteractive apt-get -qq update
+fi
+if [ "${DISTRO}" = "ubuntu" ]; then
+   sudo yum -y install lmod
+fi
 
 sudo sed -i -e '1,$s!/etc/lmod/modules!/etc/lmod/modules/Linux\n/etc/lmod/modules/ROCm\n/etc/lmod/modules/ROCmPlus\n/etc/lmod/modules/ROCmPlus-MPI\n/etc/lmod/modules/ROCmPlus-AMDResearchTools\n/etc/lmod/modules/ROCmPlus-LatestCompilers\n/etc/lmod/modules/ROCmPlus-AI\n/etc/lmod/modules/misc!' /etc/lmod/modulespath
 cat /etc/lmod/modulespath
@@ -47,4 +51,6 @@ else
   sudo ln -s /usr/share/lmod/6.6/init/cshrc /etc/profile.d/z00_lmod.csh
 fi
 
-sudo DEBIAN_FRONTEND=noninteractive apt-get -q clean && sudo rm -rf /var/lib/apt/lists/*
+if [ "${DISTRO}" = "ubuntu" ]; then
+   sudo DEBIAN_FRONTEND=noninteractive apt-get -q clean && sudo rm -rf /var/lib/apt/lists/*
+fi
